implemented basic DHCP support
DHCP is enabled on an interface when a corresponding dhcp.client.interfaces node
exists. If static addresses are configured on the same interface, they will
be interpreted as fallback addresses.

The ietf-dhcp@2016-08-25.yang used as a data model is only a draft and
has obvious shortcomings.
Also it defines client-ids and lease-time settings for DHCP clients which
we ignore currently.